huc: fix sizeof for local variables
SYMBOL::offset doubled as size in globals, but in locals it was
actually used as the offset, leaving no room to store a size anywhere
and causing sizeof to report nonsensical values. This change separates
size and offset.

(This does not currently work for either global or local structs, the
sizes of which need to be looked up in tag_table[].)

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>