huc: stmt.c: keep stack in order in for loop
Block-local variables cause the stack to grow. This happens inside the
block, which is entered repeatedly in a for loop, causing the stack
to grow with each iteration.
Because it is not easy to move the stack adjustment into the loop header,
this fix instead cleans up the stack at the end of each iteration.
Suboptimal, but works.
Part of fix for 20050826-2.c.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>