huc: stmt.c: clean up stack when returning from inside a block (with test)
doreturn() emits a branch to the exit label, but didn't take into account
that the epilog only cleans up the top-level stack, but not things
allocated in inner blocks.
This fix keeps remembers the stack pointer before entering the first
inner block and allows doreturn() to determine if any additional cleanup
is necessary. "break" may need a similar fix.
Final part of the fix for 20050826-2.c, which I will allow myself to call
one BITCH of a test case!

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>