huc: handle comma operator in parentheses correctly
Comma operator is handled in expression(), but parenthesized expressions
are handled by primary(), which calls heir1() and thus bypasses
expression(). To fix this a new function expression_ex() has been
introduced that allows to discard results of the left-hand side of a
comma operator expression, and all heir*() functions had to be modified
to pass through the comma parameter from expression() to allow primary()
to call expression_ex() correctly.

Fixes 930123-1.c.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>