huc: function return type support
This encodes the parsed return type in the function symbol, including
pointer order and structure tag if applicable. This information is
copied to the lval on return.
This commit also eliminates dead code in newfunc() that is only run
if there is no explicit return type.
Fixes 20030606-1.c, the last failing test. Hooray!

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>