tgemu: implement screen output verification
Illegal opcode 0x33 triggers a comparison of the screen contents with
an existing BMP file and returns failure if the data is not identical.
If there is no BMP file, one will be created from the screen data. (The
emulator will still fail, just to be on the safe side.)

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>