startup.asm: better support for user interrupt handlers
vsync/hsync_hook are not easy to use for custom interrupt handlers because
the standard IRQ handlers depends on them when running from CD and won't
be executed anymore as soon as a custom interrupt handler is installed.

This adds two new hooks (user_{v,h}sync_hook) that work the same way on
all media.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>