huc: fix multi-dimensional array indexing, part 2 (with test)
This is mostly a result of trial and error. The old code mapped all
pointer arrays to CINT, which of course breaks dereferencing char*
arrays. The rest of the commit serves to clean up the fallout that is
caused by not doing that.
The test case is minimal, but getting it wrong breaks tons of other tests
anyway.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>