huc: add interrupt assembler definitions, __sirq
This produces HAVE_IRQ and HAVE_SIRQ definitions in the header if one or
more functions are declared with __irq, __firq, or __sirq. This it to
inform the startup code on whether context saving is required, and how
much.
Works around well-known assembler limitation that code size may not change
between passes by adding a line of spaces after the regular header that is
overwritten with the above definitions at the end.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>