huc: implement __mapcall attribute
This creates a stub before every call to a user function that maps the
call bank where it belongs (0x8000) and restores the previous mapping
afterwards. This is useful for functions that use page 5 (0x8000) as an
extension of the data bank (0x6000) to get a continuous 16k of data.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>