huc: library.asm: add mem_mapdatabanks()
mem_mapdatabanks() maps two consecutive banks to pages 4 and 5 (0x6000
and 0x8000) to get a continuous 16k data mapping. Page 5 is normally used
as the call bank and contains trampolines for far calls, so if the code
does anything that might turn out to be a far call, it must be ensured
that either the original bank mapping is restored before the call, or
the caller has the __mapcall attribute that causes call bank mapping stubs
to be inserted where necessary.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>