huc: handle constants in global constant initializers better (i.e. at all)
The const code passes the initializer almost verbatim to the assembler,
only prefixing identifiers with an underscore. This is _always_ wrong
because the assembler interprets symbol names as addresses (breaking
scalars), and doesn't know what to do with an "&" (breaking addresses).

This sort of works, but (hopefully) doesn't allow pointer arithmetic.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>