mkit: as: allow instructions in column 0
This change allows instructions, pseudo-instructions, and macros to be
put at the start of the line.

The assembler used to treat everything without leading whitespace as a
label, so this may break legacy code if it uses labels that are named
like instructions. But then anybody calling a label "lda" deserves to be
kicked anyway.

Signed-off-by: Ulrich Hecht <ulrich.hecht@gmail.com>